Update Fraction.java
Added main method to run as program and for test and debug purposes. Added set method that allows user to set numerator and denominator at same time. Added get method that allows user to retrieve fraction expressed as a percentage. Changed equals method in order to test for equality of value not just for matching numerators and denominators.